Removed zod from bundle
