+++ packages/client/src/views/header.ts
@@ -4,7 +4,7 @@ import { env, param, updateDecoratorParams } from "../params";
 import cls from "../styles/header.module.css";
 import { defineCustomElement } from "./custom-elements";
 import { refreshAuthData } from "../helpers/auth";
-import { ClientParams } from "decorator-shared/params";
+import { type ClientParams } from "decorator-shared/params";
 import { CustomEvents } from "../events";
 
 const msgSafetyCheck = (message: MessageEvent) => {
