+++ packages/shared/params.ts
@@ -17,15 +17,6 @@ export const languageSchema = z.enum([
     "uk",
     "ru",
 ]);
-export const languageLabels = {
-    nb: "Norsk (bokmål)",
-    nn: "Norsk (nynorsk)",
-    en: "English",
-    se: "Sámegiel (samisk)",
-    pl: "Polski (polsk)",
-    uk: "Українська (ukrainsk)",
-    ru: "Русский (russisk)",
-};
 export type Language = z.infer<typeof languageSchema>;
 
 const availableLanguageSchema = z.discriminatedUnion("handleInApp", [
